correcting bug 3605899: creating big NetCDF files (> 4 GB). Tested OK on CentOS 5.8 64b by AC.

+++ src/ncdf_var_cl.cpp
@@ -962,7 +962,7 @@ else if(var_type == NC_LONG)
       }
     }
 
-    int total = 1;
+    long total = 1;
 	  
     if (e->GetKW(0) != NULL)
     {
